qr-generator/requirements.txt
azure-functions
qrcode
Pillow